backend/requirements.txt
Django>=4.2.7,<4.3.0
djangorestframework>=3.14.0,<3.15.0
python-decouple>=3.8,<4.0
django-cors-headers>=4.3.0,<4.4.0
requests>=2.31.0,<2.32.0
whitenoise>=6.5.0,<6.7.0
gunicorn>=21.2.0,<22.0.0
dj-database-url>=2.1.0,<2.2.0
Pillow>=9.5.0,<10.0.0
psycopg2-binary>=2.9.9,<3.0.0
django-oauth-toolkit>=2.3.0,<2.4.0
django-ipware>=5.0.0,<6.0.0
django-filter>=23.3,<24.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3
six>=1.16.0,<2.0.0
sqlparse>=0.4.4,<0.5.0
tzdata>=2023.3
